I think I found the issue!!!
In check collision method in ObjectManager, the for loop is calling platform.size and is increasing and adding each increasing one to arrayList, which already contains two platforms (I think) and is doing GameObject p2 thing. AddPlatform is supposed to add the two platforms though, and the checkCollision method needs the for loop for the checking collisions method with syso hi. So check next week!!