Yay square now has hitbox acting before square lands on platform
-To make hitbox land on platform before square, added 5 to tempY, which
is y variable of hitbox (because adding y lowers the value). So now
hitbox is landing on platform and checking if landing on something
before square does, and then since hitbox sinks through the object, the
square stops. Next goal is to ensure that the square does not land in
the middle.
-Previously did tempY-=gravity, tempY-=1, so since tempY was one value
larger, and is adding gravity, the gravities cancelled each other out
and raised the square to the top after it landed in the middle. Goal is
to make sure square does not land in the middle, and if it does, move
square up to top like before.