Work was all in Player class today
Recheck work in else statement in Player class. The if statement
checking if playerCollision=true is supposed to prevent the square from
stopping in the middle. It moves the square up with tempY-=1 (try doing
y-=1) but then the square moves up, falls back down when there is no
collision, collides again, and falls back down, so it is repeatedly
bouncing. Change that
Today I fixed the bouncing but it reappeared with the new
if(playerCollision==true). I fixed it with tempY-=1;, tempY-=gravity
(gravity is cancelled out, so it just stopped on the platform), and
tempY+=1, which cancelled out the tempY-=1 so it stopped on the
platform.