Today fixed the bouncing square thing
To fix bouncing square bug and sticking in middle of platform, went
into checkCollision method in ObjectManager class and did setters and
getters for x and y and checked y (look again to review).
Right now, think about making game so that there are set platforms
which remain still, and there is a separate method in ObjectManager
class that creates randomPlatforms with a timer, so that they flash and
appear randomly on the screen for only a certain amount of time (search
up ways to use random) and think about setting scores.