Add files via upload
Fixed the previous broken code Previously, the box/player was not moving. Went back to ObjectManager
and fixed the randomPlatforms method. Added a randomPlatform class,
which needs to not have any parameters in the constructor, but is
currently not working because it extends GameObject class, which holds
ints in the constructor. But the randomPlatform class should not hold
any parameters because it has randomly assigned x, y, width, and
height. Also work on incorporating the Timer into the platforms so they
appear randomly on the screen.