Made some progress today?
Fixed a floating issue with the way the square moved on the screen by
going into MyGame class and creating if statement for the square's jump
to only work if it was on the box
Still need to work on randomPlatform class!! The random platform class
does not need anything in constructor because it is not telling to
input any set values for each rectangle. But errors are created because
it extends GameObject, which takes in x, y, width, height in
constructor. Find out a way to work around that
Additionally, my addPlatform class isn't working in ObjectManager. I
don't know when it stopped working, but the syso is not printing
anymore. I deleted a Graphics g from what the method addPlatform takes
in, so that it only takes in a GameObject. I also tried to change the
draw method so that it does not take in a gameObject, and the middle
screen did not appear, since the draw needs objects to well draw.
Recheck the for statement in the addPlatform method which is commented
out to see if it makes any difference as I try to addPlatforms.
I also sometimes have a platform out of the three I added additionally
that does not appear...