Yay?
Today I quit trying to work with making randomPlatform class since it
was just a waste of time (sorta)
Instead I worked on incorporating the timer into the random platforms
so that some would appear for a selective period of time and reappear
in a different location.
It worked for a moment!
Then it broke. >:(
The hitbox stopped working after that for some reason (meaning it
stopped sticking to the square) because I’d apparently tampered around
with it, so to fix that I went into my player class and added x=tempX
and y=tempY (look again) and that fixed it
But now the remove platforms isn’t working and I see no correlation
between removing the platforms and fixing the hitbox. The original
platform, number zero in the ArrayList, is not being removed for some
reason.
Fix it.
I also changed the color to blue so it looks cool now.