huh?
I was having trouble with my remove platform methods because it would
remove platforms 1, 2, and 3 within the array list, which didn’t make
sense because the platforms being added are 0, 1, and 2, and three was
not yet added into the platforms. I tried removing 1, 2, and 4, which
did not work. So to try and fix this issue of the platforms being
incorrectly added into the array list and maybe appearing incorrectly
on the screen, I went into MyGame and moved the stopPlatforms method
(which is to stop the platforms from appearing) to the action performed
method, and moved the manager.randomplatforms below the updategamestate
and put setup level, which should not keep resetting, below the random
platforms
umm so fix the issue of the platforms removing and stop removing after
a certain point, and fix setuplevel