Hooray! Fixed removing platform issue
In actionPerformed method of MyGame class, created a setUp variable which was set to false. Then in the if statement, I asked if setUp =false, which is true, so the SetUpLevel platforms appeared. Then to have fixed platforms, I just did not remove the first couple of platforms in the arrayList (began to remove starting with platform number 2 in the ArrayList)

Work on the aesthetics next week, and if square falls off screen, then current_State switches to END_STATE. Set a limited time for a person to play the game. Ask someone else to play my game after finishing aesthetics of it! :)