HoRRah!
Nearly finished with my game. Fixed main issues, added a timer, and
when square falls off, it switches to EndState.

Problem: I think the timer starts when the screen is not the MenuState,
so when the person hits space for instructions, the instant the
InstructionState is shown, the timer starts.
-Work on making the timer start only if it is the GameState
-Work on aesthetics(yay!)
-Work on making timer visible to the player
-Fix any other issues, minor or major, that come up
HoRRah!