Tried fixing the fact that platforms keep appearing on or near the top/goal platform
Changed the y value location of each random platform as a variable
randPlatY so that I can restrict the randomly appearing platforms not
to appear near the top platform.
It’s n o t working
The randPlatY is supposed to only take in random values within the
range of right below the top platform to above the player. Figure out
next week what is wrong, because the platforms continuously appear in
the platform, beside it, or right above it.